dr11k

+++ py/SGA/coadds.py
@@ -341,9 +341,10 @@ def get_ccds(survey, ra, dec, width_pixels, pixscale=PIXSCALE, bands=BANDS):
     return ccds
 
 
-def custom_cutouts(obj, galaxy, output_dir, width, layer, pixscale=0.262,
-                   unwise_pixscale=UNWISE_PIXSCALE, galex_pixscale=GALEX_PIXSCALE,
-                   bands=GRIZ, galex=False, unwise=False):
+def custom_cutouts(obj, galaxy, output_dir, width, layer, survey, ccds=None,
+                   pixscale=0.262, unwise_pixscale=UNWISE_PIXSCALE,
+                   galex_pixscale=GALEX_PIXSCALE, bands=GRIZ, galex=False,
+                   unwise=False, ivar_cutouts=False, cleanup=True):
     """
     SGA2025_08089m6975-ccds.fits
     SGA2025_08089m6975-image.jpg
@@ -359,12 +360,23 @@ def custom_cutouts(obj, galaxy, output_dir, width, layer, pixscale=0.262,
     """
     import fitsio
     import shutil
+    from astrometry.util.util import Tan
+    from tractor.tractortime import TAITime
+    from legacypipe.bits import REF_MAP_BITS, maskbits_type
+    from legacypipe.reference import get_reference_sources, get_reference_map
+    from legacypipe.survey import LegacySurveyWcs
     from SGA.io import make_header, VEGA2AB
     from SGA.cutouts import cutout_one
 
+    # CCDs file
+    if ccds:
+        ccdsfile = os.path.join(output_dir, f'{galaxy}-ccds.fits')
+        ccds.writeto(ccdsfile, extname='CCDS', clobber=True)
+        log.info(f'Wrote {ccdsfile}')
+
     dry_run = False
     fits_cutouts = True
-    ivar_cutouts = False
+    ivar_cutouts = ivar_cutouts
     unwise_cutouts = unwise # False
     galex_cutouts = galex # False
 
@@ -393,6 +405,8 @@ def custom_cutouts(obj, galaxy, output_dir, width, layer, pixscale=0.262,
         allbands += [['FUV', 'NUV', ], ]
         allpixscale += [[GALEX_PIXSCALE]*2, ]
 
+
+    need_maskbits = True
     for fitssuffix, jpgsuffix, outjpgsuffix, allband, allpixscale in zip(
             fitssuffixes, jpgsuffixes, outjpgsuffixes, allbands, allpixscale):
         infile = os.path.join(output_dir, f'{galaxy}{jpgsuffix}.jpeg')
@@ -403,16 +417,21 @@ def custom_cutouts(obj, galaxy, output_dir, width, layer, pixscale=0.262,
         fitsfile = os.path.join(output_dir, f'{galaxy}{fitssuffix}.fits')
         try:
             imgs, hdr = fitsio.read(fitsfile, header=True)
+            for key in ['VERSION', 'BANDS', 'COMMENT']:
+                hdr.delete(key)
+            for iband in range(len(bands)):
+                hdr.delete(f'BAND{iband}')
+
+            if ivar_cutouts:
+                ivars, ivarhdr = fitsio.read(fitsfile, ext=1, header=True)
+                for key in ['VERSION', 'BANDS', 'COMMENT']:
+                    ivarhdr.delete(key)
+                for iband in range(len(bands)):
+                    ivarhdr.delete(f'BAND{iband}')
         except:
             msg = f'There was a problem reading {fitsfile} ({obj["OBJNAME"]})'
             log.critical(msg)
             return 0
-
-        hdr.delete('VERSION')
-        hdr.delete('BANDS')
-        hdr.delete('COMMENT')
-        for iband in range(len(bands)):
-            hdr.delete(f'BAND{iband}')
 
         for iband, band in enumerate(allband):
             outfile = os.path.join(output_dir, f'{galaxy}-image-{band}.fits')
@@ -421,6 +440,8 @@ def custom_cutouts(obj, galaxy, output_dir, width, layer, pixscale=0.262,
             # https://www.legacysurvey.org/dr9/description/#photometry
             if band in VEGA2AB.keys():
                 imgs[iband, :, :] *= 10.**(-0.4 * VEGA2AB[band])
+                if ivar_cutouts:
+                    ivars[iband, :, :] /= (10.**(-0.4 * VEGA2AB[band]))**2.
 
             primhdr = fitsio.FITSHDR()
             primhdr['EXTEND'] = 'T'
@@ -432,19 +453,54 @@ def custom_cutouts(obj, galaxy, output_dir, width, layer, pixscale=0.262,
                 'MAGZERO': (22.5, 'Magnitude zeropoint'),
                 'BUNIT': ('nanomaggy', 'AB mag = 22.5 - 2.5*log10(nanomaggy)'),
             }
-            outhdr = make_header(hdr, keys=hdr.keys(), extra=extra, extname=f'IMAGE_{band}')
+            outhdr = make_header(hdr, keys=hdr.keys(), extra=extra)
             fitsio.write(outfile, None, header=primhdr, clobber=True)
-            fitsio.write(outfile, imgs[iband, :, :], header=outhdr)
+            fitsio.write(outfile, imgs[iband, :, :], header=outhdr, extname=f'IMAGE_{band.upper()}')
             log.info(f'Wrote {outfile}')
 
+            if ivar_cutouts:
+                outfile = os.path.join(output_dir, f'{galaxy}-invvar-{band}.fits')
+                outhdr = make_header(ivarhdr, keys=ivarhdr.keys(), extra=extra)
+                fitsio.write(outfile, None, header=primhdr, clobber=True)
+                fitsio.write(outfile, ivars[iband, :, :], header=outhdr, extname=f'INVVAR_{band.upper()}')
+                log.info(f'Wrote {outfile}')
+
+            # maskbits image
+            if need_maskbits and band in bands:
+                wcs = LegacySurveyWcs(Tan(hdr), TAITime(None, mjd=np.mean(ccds.mjd_obs)))
+
+                refstars, _ = get_reference_sources(
+                    survey, wcs.wcs, bands=bands, tycho_stars=True,
+                    gaia_stars=True, large_galaxies=False, star_clusters=False)
+                refmap = get_reference_map(wcs.wcs, refstars)
+
+                # from runbrick.stage_image_coadds
+                MASKBITS = survey.get_maskbits()
+                maskbits = np.zeros(refmap.shape, dtype=maskbits_type)
+                for key in ['BRIGHT', 'MEDIUM', 'GALAXY', 'CLUSTER', 'RESOLVED', 'MCLOUDS']:
+                    maskbits |= MASKBITS[key] * ((refmap & REF_MAP_BITS[key]) > 0)
+
+                maskbitsfile = os.path.join(output_dir, f'{galaxy}-maskbits.fits')
+
+                outhdr = make_header(hdr, keys=hdr.keys(),
+                                     extra={'PIXSCALE': (allpixscale[iband], 'pixel scale (arcsec/pixel)')})
+                for key in ['SURVEY']:
+                    outhdr.delete(key)
+                fitsio.write(maskbitsfile, None, header=primhdr, clobber=True)
+                fitsio.write(maskbitsfile, maskbits, header=outhdr, extname='MASKBITS')
+                log.info(f'Wrote {maskbitsfile}')
+
+                need_maskbits = False
+
     # cleanup...
-    cleanfiles = [f'{basefile}.fits', f'{basefile}.jpeg']
-    if unwise:
-        cleanfiles += [f'{basefile}-unwise.fits', f'{basefile}-W1W2.jpeg']
-    if galex:
-        cleanfiles += [f'{basefile}-galex.fits', f'{basefile}-galex.jpeg']
-    for cleanfile in cleanfiles:
-        os.remove(cleanfile)
+    if cleanup:
+        cleanfiles = [f'{basefile}.fits', f'{basefile}.jpeg']
+        if unwise:
+            cleanfiles += [f'{basefile}-unwise.fits', f'{basefile}-W1W2.jpeg']
+        if galex:
+            cleanfiles += [f'{basefile}-galex.fits', f'{basefile}-galex.jpeg']
+        for cleanfile in cleanfiles:
+            os.remove(cleanfile)
 
     return 1
 
@@ -452,12 +508,13 @@ def custom_cutouts(obj, galaxy, output_dir, width, layer, pixscale=0.262,
 def custom_coadds(onegal, galaxy, survey, run, radius_mosaic_arcsec,
                   release=1000, pixscale=PIXSCALE, unwise_pixscale=UNWISE_PIXSCALE,
                   galex_pixscale=GALEX_PIXSCALE, bands=GRIZ, mp=1, layer='ls-dr11',
-                  nsigma=None, nsatur=2, rgb_stretch=1.5, no_iterative=False,
+                  nsigma=None, saddle_fraction=None, saddle_min=None, nsatur=2,
+                  rgb_stretch=1.5, no_iterative=False,
                   racolumn='GROUP_RA', deccolumn='GROUP_DEC', force_psf_detection=False,
-                  fit_on_coadds=False, just_cutouts=False, use_gpu=False, ngpu=1,
-                  threads_per_gpu=8, subsky_radii=None, just_coadds=False, missing_ok=False,
-                  force=False, cleanup=True, unwise=True, galex=False, no_gaia=False,
-                  no_tycho=False, verbose=False):
+                  fit_on_coadds=False, just_cutouts=False, ivar_cutouts=False, use_gpu=False,
+                  ngpu=1, threads_per_gpu=8, subsky_radii=None, just_coadds=False,
+                  missing_ok=False, force=False, cleanup=True, unwise=True, galex=False,
+                  no_gaia=False, no_tycho=False, verbose=False):
     """Build a custom set of large-galaxy coadds.
 
     """
@@ -485,9 +542,10 @@ def custom_coadds(onegal, galaxy, survey, run, radius_mosaic_arcsec,
     # just cutouts -- no pipeline
     if just_cutouts:
         err = custom_cutouts(onegal, galaxy, survey.output_dir, width, layer,
-                             pixscale=pixscale, bands=bands, galex=galex,
-                             unwise=unwise, unwise_pixscale=unwise_pixscale,
-                             galex_pixscale=galex_pixscale)
+                             survey, ccds=ccds, pixscale=pixscale, bands=bands,
+                             galex=galex, unwise=unwise, unwise_pixscale=unwise_pixscale,
+                             galex_pixscale=galex_pixscale, ivar_cutouts=ivar_cutouts,
+                             cleanup=cleanup)
         return err, stagesuffix
 
 
@@ -503,9 +561,12 @@ def custom_coadds(onegal, galaxy, survey, run, radius_mosaic_arcsec,
 
     if rgb_stretch:
         cmdargs += f'--rgb-stretch={rgb_stretch:.2f} '
-
     if nsigma:
         cmdargs += f'--nsigma={nsigma:.0f} '
+    if saddle_fraction:
+        cmdargs += f'--saddle-fraction={saddle_fraction:.2f} '
+    if saddle_min:
+        cmdargs += f'--saddle-min={saddle_min:.1f} '
 
     #cmdargs += '--write-stage=tims --write-stage=srcs '
     cmdargs += '--write-stage=srcs '
